consolidate soft-deletion tests
